Move request-sending functionality into a common place
