+++ lib/StripeMethod.js
@@ -1,6 +1,5 @@
 'use strict';
 const utils = require('./utils');
-const makeRequest = require('./makeRequest');
 const autoPagination = require('./autoPagination');
 const makeAutoPaginationMethods = autoPagination.makeAutoPaginationMethods;
 /**
@@ -31,7 +30,7 @@ function stripeMethod(spec) {
       spec.fullPath || this.createResourcePathWithSymbols(spec.path || '')
     );
     const requestPromise = utils.callbackifyPromiseWithTimeout(
-      makeRequest(this, args, spec, {}),
+      this._makeRequest(args, spec, {}),
       callback
     );
     // Please note `spec.methodType === 'search'` is beta functionality and this
