+++ lib/autoPagination.js
@@ -1,5 +1,4 @@
 'use strict';
-const makeRequest = require('./makeRequest');
 const utils = require('./utils');
 function makeAutoPaginationMethods(self, requestArgs, spec, firstPagePromise) {
   const promiseCache = {currentPromise: null};
@@ -21,14 +20,14 @@ function makeAutoPaginationMethods(self, requestArgs, spec, firstPagePromise) {
           'Unexpected: Stripe API response does not have a well-formed `next_page` field, but `has_more` was true.'
         );
       }
-      return makeRequest(self, requestArgs, spec, {
+      return self._makeRequest(requestArgs, spec, {
         page: pageResult.next_page,
       });
     };
   } else {
     getNextPagePromise = (pageResult) => {
       const lastId = getLastId(pageResult, reverseIteration);
-      return makeRequest(self, requestArgs, spec, {
+      return self._makeRequest(requestArgs, spec, {
         [reverseIteration ? 'ending_before' : 'starting_after']: lastId,
       });
     };
