+++ lib/stripe.js
@@ -40,6 +40,7 @@ const ALLOWED_CONFIG_PROPERTIES = [
 ];
 const EventEmitter = require('events').EventEmitter;
 const StripeResource = require('./StripeResource');
+const RequestSender = require('./makeRequest');
 Stripe.StripeResource = StripeResource;
 Stripe.resources = resources;
 const {HttpClient, HttpClientResponse} = require('./net/HttpClient');
@@ -107,6 +108,10 @@ function Stripe(key, config = {}) {
   this.webhooks = require('./Webhooks');
   this._prevRequestMetrics = [];
   this._enableTelemetry = props.telemetry !== false;
+  this._requestSender = new RequestSender(
+    this,
+    Stripe.StripeResource.MAX_BUFFERED_REQUEST_METRICS
+  );
   // Expose StripeResource on the instance too
   // @ts-ignore
   this.StripeResource = Stripe.StripeResource;
@@ -162,6 +167,7 @@ Stripe.prototype = {
   _prevRequestMetrics: null,
   _emitter: null,
   _enableTelemetry: null,
+  _requestSender: null,
   /**
    * @private
    */
