+++ lib/utils.js
@@ -117,6 +117,7 @@ const utils = {
   getOptionsFromArgs: (args) => {
     const opts = {
       auth: null,
+      host: null,
       headers: {},
       settings: {},
     };
