+++ package.json
@@ -27,6 +27,7 @@
   "main": "lib/stripe.js",
   "types": "types/index.d.ts",
   "devDependencies": {
+    "@types/mocha": "^10.0.1",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "^4.33.0",
     "chai": "^4.3.6",
