+++ testUtils/index.js
@@ -9,7 +9,7 @@ require('chai').use(require('chai-as-promised'));
 const http = require('http');
 
 const CryptoProvider = require('../lib/crypto/CryptoProvider');
-const ResourceNamespace = require('../lib/ResourceNamespace').ResourceNamespace;
+const RequestSender = require('../lib/makeRequest');
 
 const testingHttpAgent = new http.Agent({keepAlive: false});
 
@@ -58,36 +58,59 @@ const utils = (module.exports = {
     return key;
   },
 
-  getSpyableStripe: (config) => {
-    // Provide a testable stripe instance
-    // That is, with mock-requests built in and hookable
-
-    const stripe = require('../lib/stripe');
-    const stripeInstance = stripe('fakeAuthToken', config);
-
-    stripeInstance.REQUESTS = [];
-
-    for (const i in stripeInstance) {
-      makeInstanceSpyable(stripeInstance, stripeInstance[i]);
-    }
-
-    function makeInstanceSpyable(stripeInstance, thisInstance) {
-      if (thisInstance instanceof stripe.StripeResource) {
-        patchRequest(stripeInstance, thisInstance);
-      } else if (thisInstance instanceof ResourceNamespace) {
-        const namespace = thisInstance;
-
-        for (const j in namespace) {
-          makeInstanceSpyable(stripeInstance, namespace[j]);
-        }
+  getMockStripe: (config, request) => {
+    class MockRequestSender extends RequestSender {
+      _request(
+        method,
+        host,
+        path,
+        data,
+        auth,
+        options = {},
+        callback,
+        requestDataProcessor = null
+      ) {
+        return request(
+          method,
+          host,
+          path,
+          data,
+          auth,
+          options,
+          callback,
+          requestDataProcessor
+        );
       }
     }
 
-    function patchRequest(stripeInstance, instance) {
-      instance._request = function(method, host, url, data, auth, options, cb) {
+    // Provide a testable stripe instance
+    // That is, with mock-requests built in and hookable
+    const stripe = require('../lib/stripe');
+    const stripeInstance = stripe('fakeAuthToken', config);
+
+    stripeInstance._requestSender = new MockRequestSender(
+      stripeInstance,
+      stripe.StripeResource.MAX_BUFFERED_REQUEST_METRICS
+    );
+
+    return stripeInstance;
+  },
+
+  getSpyableStripe: (config) => {
+    class SpyableRequestSender extends RequestSender {
+      _request(
+        method,
+        host,
+        path,
+        data,
+        auth,
+        options = {},
+        callback,
+        requestDataProcessor = null
+      ) {
         const req = (stripeInstance.LAST_REQUEST = {
           method,
-          url,
+          url: path,
           data,
           headers: options.headers || {},
           settings: options.settings || {},
@@ -101,11 +124,11 @@ const utils = (module.exports = {
 
         const handleMockRequest = (err, req) => {
           stripeInstance.REQUESTS.push(req);
-          cb.call(this, err, {});
+          callback.call(this, err, {});
         };
 
-        if (this.requestDataProcessor) {
-          this.requestDataProcessor(
+        if (requestDataProcessor) {
+          requestDataProcessor(
             method,
             data,
             options.headers,
@@ -114,8 +137,31 @@ const utils = (module.exports = {
         } else {
           handleMockRequest(null, req);
         }
-      };
+
+        return super._request(
+          method,
+          host,
+          path,
+          data,
+          auth,
+          options,
+          callback,
+          requestDataProcessor
+        );
+      }
     }
+
+    // Provide a testable stripe instance
+    // That is, with mock-requests built in and hookable
+    const stripe = require('../lib/stripe');
+    const stripeInstance = stripe('fakeAuthToken', config);
+
+    stripeInstance.REQUESTS = [];
+
+    stripeInstance._requestSender = new SpyableRequestSender(
+      stripeInstance,
+      stripe.StripeResource.MAX_BUFFERED_REQUEST_METRICS
+    );
 
     return stripeInstance;
   },
